+++ src/lib.rs
@@ -31,6 +31,7 @@ pub mod reflate;
 pub mod resnappy;
 pub mod rezstd;
 pub mod parquet;
+pub mod safetensors;
 pub mod jpg;
 pub mod jpeg;
 pub mod fixlog;